requirements/requirements.txt
addict==2.4.0
atomicwrites==1.4.1
attrs==19.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Cython==0.29.32
defusedxml==0.7.1
fast_ctc_decode==0.3.2
filelock==3.8.0
huggingface-hub==0.9.1
idna==3.3
imagecodecs==2022.2.22
imageio==2.21.2
joblib==1.1.0
jstyleson==0.0.2
lmdb==1.3.0
more-itertools==8.14.0
networkx==2.8
nibabel==4.0.2
nltk==3.7
numpy==1.23.1
opencv-python==4.5.5.64
openvino-telemetry==2022.1.1
packaging==21.3
pandas==1.1.5
parasail==1.3.3
Pillow==9.2.0
pluggy==1.0.0
progress==1.6
py==1.11.0
py-cpuinfo==8.0.0
pyclipper==1.3.0.post3
pydicom==2.3.0
pyparsing==3.0.9
pytest==4.0.1
pytest-html==1.19.0
pytest-metadata==2.0.2
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
rawpy==0.17.2
regex==2022.8.17
requests==2.28.1
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.97
Shapely==1.8.4
six==1.16.0
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2022.8.12
tokenizers==0.10.3
tqdm==4.64.1
transformers==4.16.2
typing_extensions==4.3.0
urllib3==1.26.12
